refactor(desktop): the integration suite reads the call by concern
